Add BytesRefsFromBinaryBlockLoader
