+++ x-pack/plugin/logsdb/src/main/java/org/elasticsearch/xpack/logsdb/patterntext/PatternTextBlockLoader.java
@@ -7,11 +7,9 @@
 
 package org.elasticsearch.xpack.logsdb.patterntext;
 
-import org.apache.lucene.index.BinaryDocValues;
 import org.apache.lucene.index.LeafReaderContext;
-import org.apache.lucene.util.BytesRef;
 import org.elasticsearch.index.mapper.blockloader.docvalues.BlockDocValuesReader;
-import org.elasticsearch.index.mapper.blockloader.docvalues.BytesRefsFromCustomBinaryBlockLoader;
+import org.elasticsearch.index.mapper.blockloader.docvalues.BytesRefsFromBinaryBlockLoader;
 
 import java.io.IOException;
 
@@ -31,34 +29,7 @@ public class PatternTextBlockLoader extends BlockDocValuesReader.DocValuesBlockL
     @Override
     public AllReader reader(LeafReaderContext context) throws IOException {
         var docValues = docValuesSupplier.get(context.reader());
-        if (docValues == null) {
-            return new ConstantNullsReader();
-        }
-        return new BytesRefsFromBinary(docValues);
+        return BytesRefsFromBinaryBlockLoader.createReader(docValues);
     }
 
-    /**
-     * Read BinaryDocValues with no additional structure in the BytesRefs.
-     * Each BytesRef from the doc values maps directly to a value in the block loader.
-     */
-    public static class BytesRefsFromBinary extends BytesRefsFromCustomBinaryBlockLoader.AbstractBytesRefsFromBinary {
-        public BytesRefsFromBinary(BinaryDocValues docValues) {
-            super(docValues);
-        }
-
-        @Override
-        public void read(int doc, BytesRefBuilder builder) throws IOException {
-            if (false == docValues.advanceExact(doc)) {
-                builder.appendNull();
-                return;
-            }
-            BytesRef bytes = docValues.binaryValue();
-            builder.appendBytesRef(bytes);
-        }
-
-        @Override
-        public String toString() {
-            return "BlockDocValuesReader.Bytes";
-        }
-    }
 }
